requirements.txt
Django==1.7.1
djangorestframework==2.3.14
virtualenv==1.10.1
wikiapi==1.0.2
wsgiref==0.1.2
